feat [WIP]: Added MerchantPortal::Product::CreateForm, MerchantPortal::Product::CreateForm::Populators and MerchantPortal::Product::CreateForm::Validators. Added validations for the different models. Added weight and weight_unit to Product::Variant. Added options to Product::Variant::PersonalizationField for new field_types - select_field and multi_select_field.

+++ db/migrate/20210416130143_create_product_variants.rb
@@ -16,7 +16,7 @@ class CreateProductVariants < ActiveRecord::Migration[6.1]
       t.boolean :track_inventory, default: true
       t.boolean :backorderable, default: true
       t.boolean :giftable, default: true
-      t.integer :weight
+      t.float :weight
       t.enum :weight_unit, enum_name: :weight_unit
       t.jsonb :data, default: {}
